% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_awc.R
\name{get_awc}
\alias{get_awc}
\title{Extract the AWP for a Coordinate Pair}
\usage{
get_awc(coords)
}
\arguments{
\item{coords}{a list with lon and lat giving the coordinates}
}
\value{
a single value for AWC (in mm)
}
\description{
Extract AWC for Point from Webb et al.'s data base
}
\details{
This function extracts the AWC for a given point (list of
    coordinates as lon and lat). In case there no value at point,
    the functions looks around a bit to find the next gridpoint
    that is not NA.

The data set used is the texture derived soil water holding
    capacity from Webb et al., that is freely available via
    http://daac.ornl.gov/SOILS/guides/Webb.html
}
\examples{
get_awc(list(lon = 11.74, lat = 42.94))
}
\references{
Webb, R. W., C. E. Rosenzweig, and
    E. R. Levine. 2000. Global Soil Texture and Derived
    Water-Holding Capacities (Webb et al.). Data set. Available
    on-line [http://www.daac.ornl.gov] from Oak Ridge National
    Laboratory Distributed Active Archive Center, Oak Ridge,
    Tennessee, U.S.A. doi:10.3334/ORNLDAAC/548.
}

